Made moveBy return on cond. See justification:
In hindsight, moveBy taking a condition was a bad idea. The reason is that the
only actually useful condition is collision, and using "cond" is a bad method
for handling that. You often want to know whether you've stopped moving so you
can, e.g., set animation state, so I'm making this function terminate on cond
for now, returning false.